// bookstore-webapp/src/generated/catalog-service.ts
export interface components {
    schemas: {
        Book: {
            id: string;
            title: string;
            author: string;
            isbn?: string;
            description?: string;
            price: number;
            tags?: string[];
        };
        BookInput: {
            title: string;
            author: string;
            isbn?: string;
            description?: string;
            price: number;
            tags?: string[];
        };
        Error: {
            code: number;
            message: string;
            description?: string;
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}