// bookstore-webapp/src/generated/inventory-service.ts
export interface components {
    schemas: {
        StockLevel: {
            bookId: string;
            quantityOnHand: number;
            quantityReserved: number;
        };
        Reservation: {
            id: string;
            bookId: string;
            orderId: string;
            quantity: number;
            /** @enum {string} */
            status: "held" | "released";
        };
        Error: {
            code: number;
            message: string;
            description?: string;
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}